requirements.txt
absl-py==0.6.1
astor==0.7.1
colorama==0.4.0
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
numpy==1.15.3
protobuf==3.6.1
regex==2018.8.29
ruamel.yaml==0.15.75
six==1.11.0
tensorboard==1.11.0
tensorflow-gpu==1.11.0
termcolor==1.1.0
tqdm==4.28.1
Werkzeug==0.14.1
